feat: ListCrudRepository

+++ src/main/java/com/ecommerce/engine/repository/GroupRepository.java
@@ -3,5 +3,5 @@ package com.ecommerce.engine.repository;
 import com.ecommerce.engine.repository.entity.Group;
 import org.springframework.data.repository.ListCrudRepository;
 
-public interface GroupRepository extends ListCrudRepository<Group, Long> {
+public interface GroupRepository extends ListCrudRepository<Group, Integer> {
 }
